Split location dataLoaders (by ID and by Term)
