+++ apps/graphql/src/apps/common/types/inputs/LocationTypeInput.js
@@ -1,0 +1,13 @@
+// @flow
+
+export type LocationTypeInput =
+  | 'STATION'
+  | 'AIRPORT'
+  | 'BUS_STATION'
+  | 'CITY'
+  | 'AUTONOMOUS_TERRITORY'
+  | 'SUBDIVISION'
+  | 'COUNTRY'
+  | 'REGION'
+  | 'CONTINENT'
+  | 'SPECIAL';
